use-cases/brain-mri/experiment/requirements.txt
# Requirements for the IDE

determined==0.19.0
torch==1.10.2
Pillow==9.0.1
numpy==1.22.2
torchvision==0.11.3
python_pachyderm==7.3.0
scikit-image==0.19.2
